+++ spec-toolkit/src/model/Config.ts
@@ -1,50 +1,8 @@
 import fs from "fs-extra";
 import path from "path";
-
-//////////////////////////////////////////
-// NEW CONFIG (TODO, WIP)               //
-//////////////////////////////////////////
-
-export interface SpecToolkitConfig {
-  specs?: SpecConfig[];
-  specExtensions?: ExtensionConfig[];
-  // specOutput?: {};
-  examples?: ExampleConfig[];
-}
-
-// Configuration Document Interface
-export interface SpecConfig {
-  id: string;
-  title: string;
-  sourceFile: string;
-  sourceFileIntroduction: string;
-  targetFile: string;
-  targetFolder?: string;
-  extensionTargetFile?: string;
-  sideBarPosition: number;
-  sideBarDescription: string;
-  /** List of bullet points to add at the top as quick facts / links (in markdown) */
-  facts?: string[];
-}
-
-export interface ExtensionConfig {
-  targetDocument: string;
-  targetDocumentFileName: string;
-  targetDocumentFolder: string;
-  specExtensions: string[]; // array of file paths
-}
-
-export interface ExampleConfig {
-  folder: string; // TODO: Point to folder? Or just give a glob pattern?
-}
-
-//////////////////////////////////////////
-// ORIGINAL CONFIG                      //
-//////////////////////////////////////////
-
 export interface ConfigFile {
   generalConfig: GeneralConfig;
-  docsConfig: DocsConfig[];
+  docsConfig: SpecConfig[];
 }
 
 // Configuration General Configuration Interface
@@ -52,16 +10,19 @@ export interface GeneralConfig {
   sortProperties: boolean;
 }
 
+export type SpecType = "main" | "extension";
+export type SpecConfig = MainSpecConfig | ExtensionSpecConfig;
 // Configuration Document Interface
-export interface DocsConfig {
+export interface MainSpecConfig {
+  type: "main";
   id: string;
   title: string;
-  sourceFile: string;
-  sourceFileIntroduction?: string;
+  sourceFilePath: string;
+  sourceIntroductionFilePath?: string;
   sourceFileOutro?: string;
-  targetFile: string;
+  targetMarkdownFilePath: string;
+  targetJsonSchemaFilePath: string;
   targetFolder?: string;
-  extensionTargetFile?: string;
   sideBarPosition: number;
   sideBarDescription: string;
 
@@ -74,24 +35,38 @@ export interface DocsConfig {
   /** List of bullet points to add at the top as quick facts / links (in markdown) */
   facts?: string[];
 }
+export interface ExtensionSpecConfig {
+  type: "extension";
+  id: string;
+  title: string;
+  sourceFilePath: string;
+  sourceIntroductionFilePath?: string;
+  sourceFileOutro?: string;
+  targetMarkdownFilePath: string;
+  targetJsonSchemaFilePath: string;
+  targetFolder?: string;
+  sideBarPosition: number;
+  sideBarDescription: string;
 
-//Get Config for a given ID
-export function getConfigForId(docsConfigs: DocsConfig[], id: string): DocsConfig | null {
-  for (let i = 0; i < docsConfigs.length; i++) {
-    if (docsConfigs[i].id === id) {
-      return docsConfigs[i];
-    }
-  }
-  return null;
+  /**
+   * Overrides docusaurus max MD heading level that should be displayed in the table of contents (in the docusaurus right sidebar).
+   * Docusaurus default value if not specified: 3
+   */
+  tocMaxHeadingLevel?: number;
+
+  /** List of bullet points to add at the top as quick facts / links (in markdown) */
+  facts?: string[];
+  targetDocument: string;
+  targetLink: string;
 }
 
 //Retrieve Text from Introduction File
-export function getIntroductionText(docConfig: DocsConfig): string {
-  if (!docConfig.sourceFileIntroduction) {
+export function getIntroductionText(docConfig: SpecConfig): string {
+  if (!docConfig.sourceIntroductionFilePath) {
     return "";
   }
 
-  const mdFilePath = path.resolve(docConfig.sourceFileIntroduction);
+  const mdFilePath = path.resolve(docConfig.sourceIntroductionFilePath);
 
   if (!fs.existsSync(mdFilePath)) {
     throw new Error("Could not read markdown file: " + mdFilePath);
@@ -100,7 +75,7 @@ export function getIntroductionText(docConfig: DocsConfig): string {
   return fs.readFileSync(mdFilePath, "utf-8");
 }
 //Retrieve Text from Introduction File
-export function getOutroText(docConfig: DocsConfig): string {
+export function getOutroText(docConfig: SpecConfig): string {
   if (!docConfig.sourceFileOutro) {
     return "";
   }
@@ -114,10 +89,10 @@ export function getOutroText(docConfig: DocsConfig): string {
   return fs.readFileSync(mdFilePath, "utf-8");
 }
 
-export function getTargetDocumentForDocumentId(documentID: string, docsConfigs: DocsConfig[]): string {
+export function getTargetDocumentForDocumentId(documentID: string, docsConfigs: SpecConfig[]): string {
   for (let i = 0; i < docsConfigs.length; i++) {
     if (docsConfigs[i].id === documentID) {
-      return docsConfigs[i].targetFile;
+      return docsConfigs[i].targetMarkdownFilePath;
     }
   }
   return "";